add useSettings
